+++ examples/mpi_vegas_example.cpp
@@ -91,8 +91,8 @@ int main(int argc, char* argv[])
 		MPI_COMM_WORLD
 	);
 
-	auto result = hep::cumulative_result(results.begin(), results.end());
-	double chi_square_dof = hep::chi_square_dof(results.begin(), results.end());
+	auto result = hep::cumulative_result0(results.begin(), results.end());
+	double chi_square_dof = hep::chi_square_dof0(results.begin(), results.end());
 
 	if (rank == 0)
 	{