+++ tests/test_mc_helper.cpp
@@ -9,12 +9,12 @@ typedef testing::Types<float, double, long double> MyT;
 template <typename T> class McHelper : public testing::Test { };
 TYPED_TEST_CASE(McHelper, MyT);
 
-TYPED_TEST(McHelper, CumulativeResultWithZero)
+TYPED_TEST(McHelper, CumulativeResult0WithZero)
 {
 	typedef TypeParam T;
 
 	std::vector<hep::mc_result<T>> zero_results;
-	auto result = hep::cumulative_result(zero_results.begin(), zero_results.end());
+	auto result = hep::cumulative_result0(zero_results.begin(), zero_results.end());
 
 	// assert that results are reasonable even if there is no real input
 	EXPECT_EQ( 0U  , result.calls );
@@ -22,21 +22,48 @@ TYPED_TEST(McHelper, CumulativeResultWithZero)
 	EXPECT_EQ( T() , result.error );
 }
 
-TYPED_TEST(McHelper, CumulativeResultWithOne)
+TYPED_TEST(McHelper, CumulativeResult1WithZero)
+{
+	typedef TypeParam T;
+
+	std::vector<hep::mc_result<T>> zero_results;
+	auto result = hep::cumulative_result1(zero_results.begin(), zero_results.end());
+
+	// assert that results are reasonable even if there is no real input
+	EXPECT_EQ( 0U  , result.calls );
+	EXPECT_EQ( T() , result.value );
+	EXPECT_EQ( T() , result.error );
+}
+
+TYPED_TEST(McHelper, CumulativeResult0WithOne)
 {
 	typedef TypeParam T;
 
 	std::vector<hep::mc_result<T>> one_result = {
 		hep::mc_result<T>(1000, T(10.0), T(20.0))
 	};
-	auto result = hep::cumulative_result(one_result.begin(), one_result.end());
+	auto result = hep::cumulative_result0(one_result.begin(), one_result.end());
 
 	EXPECT_EQ( one_result.front().calls , result.calls );
 	EXPECT_NEAR( one_result.front().value , result.value , T(1e-10) );
 	EXPECT_NEAR( one_result.front().error , result.error , T(1e-10) );
 }
 
-TYPED_TEST(McHelper, CumulativeResultWithTwo)
+TYPED_TEST(McHelper, CumulativeResult1WithOne)
+{
+	typedef TypeParam T;
+
+	std::vector<hep::mc_result<T>> one_result = {
+		hep::mc_result<T>(1000, T(10.0), T(20.0))
+	};
+	auto result = hep::cumulative_result1(one_result.begin(), one_result.end());
+
+	EXPECT_EQ( one_result.front().calls , result.calls );
+	EXPECT_NEAR( one_result.front().value , result.value , T(1e-10) );
+	EXPECT_NEAR( one_result.front().error , result.error , T(1e-10) );
+}
+
+TYPED_TEST(McHelper, CumulativeResult0WithTwo)
 {
 	typedef TypeParam T;
 
@@ -45,7 +72,7 @@ TYPED_TEST(McHelper, CumulativeResultWithTwo)
 		hep::mc_result<T>(100, T(100.0), T(200.0)),
 		hep::mc_result<T>(100, T(100.0), T(200.0))
 	};
-	auto result = hep::cumulative_result(two_results.begin(), two_results.end());
+	auto result = hep::cumulative_result0(two_results.begin(), two_results.end());
 
 	// should give two times the calls
 	EXPECT_EQ( 2 * two_results.front().calls , result.calls );
@@ -55,29 +82,7 @@ TYPED_TEST(McHelper, CumulativeResultWithTwo)
 	EXPECT_NEAR( two_results.front().error / std::sqrt(T(2.0)) , result.error , T(1e-10) );
 }
 
-TYPED_TEST(McHelper, ChiSquareDofWithZero)
-{
-	typedef TypeParam T;
-
-	std::vector<hep::mc_result<T>> zero_results;
-	T result = hep::chi_square_dof(zero_results.begin(), zero_results.end());
-
-	EXPECT_NEAR( T() , result , T(1e-10) );
-}
-
-TYPED_TEST(McHelper, ChiSquareDofWithOne)
-{
-	typedef TypeParam T;
-
-	std::vector<hep::mc_result<T>> one_result = {
-		hep::mc_result<T>(1000, T(10.0), T(20.0))
-	};
-	T result = hep::chi_square_dof(one_result.begin(), one_result.end());
-
-	EXPECT_EQ( std::numeric_limits<T>::infinity() , result );
-}
-
-TYPED_TEST(McHelper, ChiSquareDofWithTwo)
+TYPED_TEST(McHelper, CumulativeResult1WithTwo)
 {
 	typedef TypeParam T;
 
@@ -86,7 +91,104 @@ TYPED_TEST(McHelper, ChiSquareDofWithTwo)
 		hep::mc_result<T>(100, T(100.0), T(200.0)),
 		hep::mc_result<T>(100, T(100.0), T(200.0))
 	};
-	T result = hep::chi_square_dof(two_results.begin(), two_results.end());
+	auto result = hep::cumulative_result1(two_results.begin(), two_results.end());
+
+	// should give two times the calls
+	EXPECT_EQ( 2 * two_results.front().calls , result.calls );
+	// the same result and
+	EXPECT_NEAR( two_results.front().value , result.value , T(1e-10) );
+	// zero error
+	EXPECT_EQ( T() , result.error );
+}
+
+TYPED_TEST(McHelper, CumulativeResult1WithThree)
+{
+	typedef TypeParam T;
+
+	// two times the same result
+	std::vector<hep::mc_result<T>> two_results = {
+		hep::mc_result<T>(100, T(100.0), T(200.0)),
+		hep::mc_result<T>(100, T(100.0), T(200.0)),
+		hep::mc_result<T>(100, T(100.0), T(200.0)),
+	};
+	auto result = hep::cumulative_result1(two_results.begin(), two_results.end());
+
+	// should give three times the calls
+	EXPECT_EQ( 3 * two_results.front().calls , result.calls );
+	// the same result and
+	EXPECT_NEAR( two_results.front().value , result.value , T(1e-10) );
+	// zero error
+	EXPECT_EQ( T() , result.error );
+}
+
+TYPED_TEST(McHelper, ChiSquareDof0WithZero)
+{
+	typedef TypeParam T;
+
+	std::vector<hep::mc_result<T>> zero_results;
+	T result = hep::chi_square_dof0(zero_results.begin(), zero_results.end());
+
+	EXPECT_NEAR( T() , result , T(1e-10) );
+}
+
+TYPED_TEST(McHelper, ChiSquareDof1WithZero)
+{
+	typedef TypeParam T;
+
+	std::vector<hep::mc_result<T>> zero_results;
+	T result = hep::chi_square_dof1(zero_results.begin(), zero_results.end());
+
+	EXPECT_NEAR( T() , result , T(1e-10) );
+}
+
+TYPED_TEST(McHelper, ChiSquareDof0WithOne)
+{
+	typedef TypeParam T;
+
+	std::vector<hep::mc_result<T>> one_result = {
+		hep::mc_result<T>(1000, T(10.0), T(20.0))
+	};
+	T result = hep::chi_square_dof0(one_result.begin(), one_result.end());
+
+	EXPECT_EQ( std::numeric_limits<T>::infinity() , result );
+}
+
+TYPED_TEST(McHelper, ChiSquareDof1WithOne)
+{
+	typedef TypeParam T;
+
+	std::vector<hep::mc_result<T>> one_result = {
+		hep::mc_result<T>(1000, T(10.0), T(20.0))
+	};
+	T result = hep::chi_square_dof1(one_result.begin(), one_result.end());
+
+	EXPECT_EQ( std::numeric_limits<T>::infinity() , result );
+}
+
+TYPED_TEST(McHelper, ChiSquareDof0WithTwo)
+{
+	typedef TypeParam T;
+
+	// two times the same result
+	std::vector<hep::mc_result<T>> two_results = {
+		hep::mc_result<T>(100, T(100.0), T(200.0)),
+		hep::mc_result<T>(100, T(100.0), T(200.0))
+	};
+	T result = hep::chi_square_dof0(two_results.begin(), two_results.end());
+
+	EXPECT_NEAR( T() , result , T(1e-10) );
+}
+
+TYPED_TEST(McHelper, ChiSquareDof1WithTwo)
+{
+	typedef TypeParam T;
+
+	// two times the same result
+	std::vector<hep::mc_result<T>> two_results = {
+		hep::mc_result<T>(100, T(100.0), T(200.0)),
+		hep::mc_result<T>(100, T(100.0), T(200.0))
+	};
+	T result = hep::chi_square_dof1(two_results.begin(), two_results.end());
 
 	EXPECT_NEAR( T() , result , T(1e-10) );
 }